cleanup unused files and move react components, css and images into their own dirs
