+++ data/inspector/components/actors-toolbar.js
