+++ data/inspector/components/factories.js
