+++ data/inspector/components/packet-details.js
