+++ data/inspector/components/packets-summary.js
