+++ data/inspector/components/packets-toolbar.js
