+++ data/inspector/components/pools.js
