+++ data/inspector/components/search-box.js
