+++ data/inspector/css/actors-panel.css
