+++ data/inspector/css/base.css
