+++ data/inspector/css/packets-panel.css
