+++ data/inspector/css/search-box.css
