+++ data/inspector/css/toolbar.css
