+++ data/inspector/css/toolbox.css
