+++ data/inspector/main.html
@@ -9,12 +9,12 @@
     <link href="../lib/bootstrap/css/bootstrap.min.css" rel="stylesheet">
 
     <!-- Custom styles -->
-    <link href="base.css" rel="stylesheet">
-    <link href="toolbox.css" rel="stylesheet">
-    <link href="toolbar.css" rel="stylesheet">
-    <link href="packets-panel.css" rel="stylesheet">
-    <link href="actors-panel.css" rel="stylesheet">
-    <link href="search-box.css" rel="stylesheet">
+    <link href="css/base.css" rel="stylesheet">
+    <link href="css/toolbox.css" rel="stylesheet">
+    <link href="css/toolbar.css" rel="stylesheet">
+    <link href="css/packets-panel.css" rel="stylesheet">
+    <link href="css/actors-panel.css" rel="stylesheet">
+    <link href="css/search-box.css" rel="stylesheet">
     <link href="chrome://firebug.sdk/skin/domTree.css" rel="stylesheet">
 
     <!-- Load RequireJS config file as the entry point module -->
