--- data/inspector/packet-editor.css
@@ -1,6 +1,0 @@
-/* See license.txt for terms of usage */
-
-body {
-  padding: 10px;
-  background-color: white;
-}
