--- data/inspector/packet-editor.js
@@ -1,5 +1,0 @@
-/* See license.txt for terms of usage */
-
-define(function(require) {
-
-});
