+++ data/inspector/res/icon-16.png
